+++ docs/examples/python/template-tag-bad-view.py
@@ -2,5 +2,8 @@ from django.shortcuts import render
 
 
 def example_view(request):
-    context_vars = {"my_variable": "example_project.my_app.components.hello_world"}
-    return render(request, "my-template.html", context_vars)
+    return render(
+        request,
+        "my_template.html",
+        context={"my_variable": "example_project.my_app.components.hello_world"},
+    )