+++ tests/test_app/pyscript/views.py
@@ -1,0 +1,5 @@
+from django.shortcuts import render
+
+
+def pyscript(request, path=None):
+    return render(request, "pyscript.html", {})